requirement.txt
Django==3.2.9
mysqlclient==2.1.0
factory_boy
django-environ